pyproject.toml
# pyproject.toml

[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "inuits_policy_based_auth"
version = "4.1.0"
description = "Module for securing API endpoints based on policies."
readme = "README.md"
authors = [{ name = "Inuits", email = "developers@inuits.eu" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
keywords = ["inuits", "policy", "auth"]
dependencies = [
    "Authlib>=1.2.0",
    "cffi>=1.15.1",
    "click>=8.1.3",
    "cryptography>=40.0.2",
    "MarkupSafe>=2.1.2",
    "pycparser>=2.21",
    "requests>=2.28.2",
    "Werkzeug>=2.2.3",
]

[project.optional-dependencies]
dev = [
    "black==23.3.0",
    "build==0.10.0",
    "coverage==7.2.7",
    "Flask-RESTful==0.3.10",
    "Flask==2.3.2",
    "mypy-extensions==1.0.0",
    "pathspec==0.11.1",
    "platformdirs==3.6.0",
    "pyjwt==2.7.0",
    "pytest==7.3.2",
    "python-dotenv==1.0.0",
    "tomli==2.0.1",
    "twine==4.0.2",
]

[project.urls]
Homepage = "https://github.com/inuits/inuits-policy-based-auth"
